updates forgetpass

+++ src/main/java/com/digimenu/main/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.digimenu.main.security.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>
+public interface UserRepository extends JpaRepository<User, Long>
 {
     User findByUsername(String username);
     User findByEmail(String email);
